adding check for SA authentication step

+++ scripts/bulk_set_requester_pays.sh
@@ -65,7 +65,9 @@ done < <(cut -d "," -f1 $CSVIN | tail -n +2)
 # get key
 docker run -v $HOME:/root --rm -it broadinstitute/dsde-toolbox vault read --format=json secret/dsde/prod/common/requester-pays.json | jq .data > rp-key.json
 # authenticate as service account
-gcloud auth activate-service-account ${SERVICE_ACCOUNT} --project="broad-dsde-prod" --key-file=rp-key.json
+if ! gcloud auth activate-service-account ${SERVICE_ACCOUNT} --project="broad-dsde-prod" --key-file=rp-key.json; then
+  echo "Error authenticating service account."
+  exit 1
 rm rp-key.json
 
 # wait for IAM changes to propagate
